backend/requirements.txt
Flask == 3.0.0
Flask-Cors == 4.0.0
pymongo == 4.5.0
dnspython == 2.4.2
